Non-working bindings for actions + fix circular dependency hell
