+++ packages/toolpad-app/src/toolpad/AppEditor/PageEditor/QueryEditor/QueryEditorDialog.tsx
@@ -30,6 +30,7 @@ import useBoolean from '../../../../utils/useBoolean';
 import { useNodeNameValidation } from '../../HierarchyExplorer/validation';
 import useEvent from '../../../../utils/useEvent';
 import useUnsavedChangesConfirm from '../../../hooks/useUnsavedChangesConfirm';
+import { PropBindingEditorDialog } from '../../BindingEditor/PropBindingEditorDialog';
 
 interface QueryEditorDialogActionsProps {
   saveDisabled?: boolean;
@@ -322,7 +323,7 @@ export default function QueryNodeEditorDialog<Q>({
                 </MenuItem>
                 <MenuItem value="mutation">Only fetch on manual action</MenuItem>
               </TextField>
-              <BindableEditor
+              <BindableEditor<boolean>
                 liveBinding={liveEnabled}
                 globalScope={pageState}
                 globalScopeMeta={globalScopeMeta}
@@ -332,6 +333,7 @@ export default function QueryNodeEditorDialog<Q>({
                 value={input.attributes.enabled ?? appDom.createConst(true)}
                 onChange={handleEnabledChange}
                 disabled={mode !== 'query'}
+                Dialog={PropBindingEditorDialog}
               />
               <TextField
                 InputProps={{
