+++ packages/toolpad-app/src/toolpadDataSources/local/client.tsx
@@ -25,6 +25,7 @@ import { usePrivateQuery } from '../context';
 import BindableEditor from '../../toolpad/AppEditor/PageEditor/BindableEditor';
 import { getDefaultControl } from '../../toolpad/propertyControls';
 import { errorFrom } from '../../utils/errors';
+import { PropBindingEditorDialog } from '../../toolpad/AppEditor/BindingEditor/PropBindingEditorDialog';
 
 const EMPTY_PARAMS: BindableAttrEntries = [];
 
@@ -155,6 +156,7 @@ function QueryEditor({
                       params: newParams,
                     }));
                   }}
+                  Dialog={PropBindingEditorDialog}
                 />
               ) : null;
             })}
