+++ packages/toolpad-app/src/toolpadDataSources/rest/BodyEditor.tsx
@@ -26,6 +26,7 @@ import * as appDom from '../../appDom';
 import TabPanel from '../../components/TabPanel';
 import ParametersEditor from '../../toolpad/AppEditor/PageEditor/ParametersEditor';
 import { HTTP_NO_BODY } from './shared';
+import { PropBindingEditorDialog } from '../../toolpad/AppEditor/BindingEditor/PropBindingEditorDialog';
 
 interface ContentTypeSpec {
   alias: string;
@@ -136,7 +137,7 @@ function RawBodyEditor({
           </TextField>
         ),
       })}
-      <BindableEditor
+      <BindableEditor<string>
         sx={{ mt: 1 }}
         liveBinding={liveContent}
         globalScope={globalScope}
@@ -156,6 +157,7 @@ function RawBodyEditor({
         onChange={handleValueChange}
         label="json"
         disabled={disabled}
+        Dialog={PropBindingEditorDialog}
       />
     </React.Fragment>
   );
